fixing jest test
