+++ app/client/src/entities/Engine/AppEditorEngine.ts
@@ -59,7 +59,7 @@ import {
 import { getFirstTimeUserOnboardingComplete } from "selectors/onboardingSelectors";
 import { isAirgapped } from "@appsmith/utils/airgapHelpers";
 import { getAIPromptTriggered } from "utils/storage";
-import { trackOpenEditorTabs } from "../../utils/editor/EditorUtils";
+import { trackOpenEditorTabs } from "../../utils/editor/browserTabsTracking";
 
 export default class AppEditorEngine extends AppEngine {
   constructor(mode: APP_MODE) {
